app/requirements.txt
Django==3.2.10
django-api==0.1.3
django-filter==21.1
django-rest==0.8.7
django-sslserver==0.22
djangorestframework==3.12.4
drf-jwt==1.19.1
drf-yasg==1.20.0
pytz==2021.3
psycopg2==2.8.6
psycopg2==2.8.6
api==0.0.7
Pillow==8.3.2